chore(geo): refactor
