+++ packages/geo/src/index.ts
@@ -1,2 +1,3 @@
 export * from './compute-markers.ts'
-export * from './geo-utils.ts'
+export * from './geo-json.ts'
+export * from './geo-utilities.ts'
